feat(homepage): move teacher homepage to root url
