+++ src/__tests__/pages/index.test.tsx
@@ -1,71 +1,85 @@
-import { getByRole, screen, within } from "@testing-library/react";
+import { screen, within, getByRole, fireEvent } from "@testing-library/react";
 
-import Home, {
-  getStaticProps,
+import Teachers, {
   HomePageProps,
   SerializedPost,
-} from "../../pages";
-import CMSClient from "../../node-lib/cms";
-import {
-  BlogPostPreview,
-  HomePage,
-  WebinarPreview,
-} from "../../common-lib/cms-types";
-import renderWithProviders from "../__helpers__/renderWithProviders";
-import renderWithSeo from "../__helpers__/renderWithSeo";
-import { mockSeo, portableTextFromString } from "../__helpers__/cms";
+  getStaticProps,
+  TeachersHomePageProps,
+} from "@/pages/index";
+import CMSClient from "@/node-lib/cms";
+import { BlogPostPreview, WebinarPreview } from "@/common-lib/cms-types";
+import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
+import keyStageKeypad from "@/browser-lib/fixtures/keyStageKeypad";
 
-jest.mock("../../node-lib/cms");
-
-const render = renderWithProviders();
+jest.mock("src/node-lib/cms");
 
 const mockCMSClient = CMSClient as jest.MockedObject<typeof CMSClient>;
-
-const pageData = {
-  id: "homepage",
-  heading: "Oak",
-  summaryPortableText: portableTextFromString("Here's the page summary"),
-  notification: {
-    enabled: true,
-    heading: "Read this news!",
-    link: {
-      linkType: "external",
-      external: "https://example.com",
-    },
+const mockPosts = [
+  {
+    id: "1",
+    type: "blog-post",
+    title: "Some blog post",
+    slug: "some-blog-post",
+    date: new Date("2021-12-01").toISOString(),
+    category: { title: "Some category", slug: "some-category" },
   },
-} as unknown as HomePage;
+  {
+    id: "2",
+    type: "blog-post",
+    title: "Some other post",
+    slug: "some-other-post",
+    date: new Date("2021-12-01").toISOString(),
+    category: { title: "Some category", slug: "some-category" },
+  },
+] as SerializedPost[];
+const props: TeachersHomePageProps = {
+  pageData: {
+    heading: "",
+    id: "",
+    summaryPortableText: [],
+    notification: { enabled: false },
+    sidebarCard1: { title: "", bodyPortableText: [] },
+    sidebarCard2: { title: "", bodyPortableText: [] },
+  },
+  posts: mockPosts,
+  curriculumData: {
+    keyStages: keyStageKeypad.keyStages,
+  },
+};
 
 jest.mock("next/dist/client/router", () => require("next-router-mock"));
 
+const render = renderWithProviders();
+
 describe("pages/index.tsx", () => {
-  it("Renders correct title and summary", () => {
-    render(<Home pageData={pageData} posts={[]} />);
+  it("Renders correct title ", () => {
+    render(<Teachers {...props} />);
 
     const h1 = screen.getByRole("heading", { level: 1 });
-    expect(h1).toHaveTextContent("Oak");
-
-    const firstH2 = screen.getAllByRole("heading", { level: 2 })[0];
-    expect(firstH2).toHaveTextContent("Here's the page summary");
+    expect(h1).toHaveTextContent("Teachers");
   });
+  it("Render correct tab after selecting tab", () => {
+    const { getByTitle, getAllByTitle } = render(<Teachers {...props} />);
+    const curriculumPlans = getAllByTitle("Curriculum plans");
+    const curriculumPlansButton = curriculumPlans[1];
+    if (curriculumPlansButton) {
+      fireEvent.click(curriculumPlansButton);
+      const curriculumH1 = screen.getByRole("heading", { level: 1 });
+      expect(curriculumH1).toHaveTextContent("Teachers & subject leads");
+    } else {
+      throw new Error("Could not find curriculum plans button element");
+    }
 
-  it("Renders the notification when enabled", () => {
-    render(<Home pageData={pageData} posts={[]} />);
+    fireEvent.click(getByTitle("Pupils"));
+    const pupilsH1 = screen.getByRole("heading", { level: 1 });
+    expect(pupilsH1).toHaveTextContent("Pupils");
 
-    expect(screen.queryByText("Read this news!")).toBeInTheDocument();
+    fireEvent.click(getByTitle("Teaching resources"));
+    const teachersH1 = screen.getByRole("heading", { level: 1 });
+    expect(teachersH1).toHaveTextContent("Teachers");
   });
-
-  it("Does not render the notification when disabled", () => {
-    const disabledNotificationPageData = {
-      ...pageData,
-      notification: { ...pageData.notification, enabled: false },
-    } as HomePage;
-    render(<Home pageData={disabledNotificationPageData} posts={[]} />);
-
-    expect(screen.queryByText("Read this news!")).not.toBeInTheDocument();
-  });
-
   it("Renders a link to the blog list", () => {
-    render(<Home pageData={pageData} posts={[]} />);
+    render(<Teachers {...props} />);
 
     const blogLink = screen.getByText("All blogs");
     expect(blogLink).toBeInTheDocument();
@@ -73,26 +87,7 @@ describe("pages/index.tsx", () => {
   });
 
   it("Renders the provided blog posts", async () => {
-    const mockPosts = [
-      {
-        id: "1",
-        type: "blog-post",
-        title: "Some blog post",
-        slug: "some-blog-post",
-        date: new Date("2021-12-01").toISOString(),
-        category: { title: "Some category", slug: "some-category" },
-      },
-      {
-        id: "2",
-        type: "blog-post",
-        title: "Some other post",
-        slug: "some-other-post",
-        date: new Date("2021-12-01").toISOString(),
-        category: { title: "Some category", slug: "some-category" },
-      },
-    ] as SerializedPost[];
-
-    render(<Home pageData={pageData} posts={mockPosts} />);
+    render(<Teachers {...props} />);
 
     const list = screen
       .getAllByRole("list")
@@ -110,24 +105,6 @@ describe("pages/index.tsx", () => {
         name: "Some blog post",
       }),
     ).toHaveAttribute("href", "/blog/some-blog-post");
-  });
-
-  describe.skip("SEO", () => {
-    it("renders the correct SEO details", () => {
-      const { seo } = renderWithSeo()(
-        <Home pageData={{ ...pageData, seo: undefined }} posts={[]} />,
-      );
-
-      expect(seo).toEqual({});
-    });
-
-    it("renders the correct SEO details from the CMS", () => {
-      const { seo } = renderWithSeo()(
-        <Home pageData={{ ...pageData, seo: mockSeo() }} posts={[]} />,
-      );
-
-      expect(seo).toEqual({});
-    });
   });
 
   describe("getStaticProps", () => {
@@ -159,7 +136,7 @@ describe("pages/index.tsx", () => {
       jest.clearAllMocks();
       jest.resetModules();
 
-      mockCMSClient.homepage.mockResolvedValue(pageData);
+      mockCMSClient.homepage.mockResolvedValue(props.pageData);
       mockCMSClient.blogPosts.mockResolvedValue([]);
       mockCMSClient.webinars.mockResolvedValue([]);
     });
@@ -173,7 +150,9 @@ describe("pages/index.tsx", () => {
         mockPost,
         mockPost2,
       ]);
-      const result = (await getStaticProps({})) as { props: HomePageProps };
+      const result = (await getStaticProps({
+        params: {},
+      })) as { props: HomePageProps };
 
       expect(result.props?.posts).toHaveLength(4);
     });
@@ -184,7 +163,9 @@ describe("pages/index.tsx", () => {
         { ...mockPost, id: "3", date: new Date("2021-01-01") },
         { ...mockPost, id: "1", date: new Date("2023-01-01") },
       ]);
-      const result = (await getStaticProps({})) as { props: HomePageProps };
+      const result = (await getStaticProps({
+        params: {},
+      })) as { props: HomePageProps };
 
       const postIds = result.props.posts.map((p) => p.id);
       expect(postIds).toEqual(["1", "2", "3"]);
@@ -196,15 +177,19 @@ describe("pages/index.tsx", () => {
         { ...mockPost3, id: "3", date: new Date("2021-01-01") },
         { ...mockPost3, id: "1", date: new Date("4023-01-01") },
       ]);
-      const result = (await getStaticProps({})) as { props: HomePageProps };
+      const result = (await getStaticProps({
+        params: {},
+      })) as { props: HomePageProps };
 
-      const postIds = result.props.posts.map((p) => p.id);
+      const postIds = result.props.posts.map((p) => p.id as string);
       expect(postIds).toEqual(["2", "3"]);
     });
 
     it("Should not fetch draft content by default", async () => {
       mockCMSClient.blogPosts.mockResolvedValueOnce([mockPost]);
-      await getStaticProps({});
+      await getStaticProps({
+        params: {},
+      });
 
       expect(mockCMSClient.blogPosts).toHaveBeenCalledWith(
         expect.objectContaining({
