+++ src/node-lib/curriculum-api-2023/__mocks__/index.ts
@@ -2,10 +2,14 @@ import { CurriculumApi } from "@/node-lib/curriculum-api-2023";
 import subjectPhaseOptionsFixture from "@/node-lib/curriculum-api-2023/fixtures/subjectPhaseOptions.fixture";
 import { curriculumOverviewMVFixture } from "@/node-lib/curriculum-api-2023/fixtures/curriculumOverview.fixture";
 import { subjectListingFixture2023 } from "@/node-lib/curriculum-api/fixtures/subjectListing.fixture";
+import teachersHomePageFixture from "@/node-lib/curriculum-api/fixtures/teachersHomePage.fixture";
 
 const curriculumApi: Pick<
   CurriculumApi,
-  "subjectPhaseOptions" | "curriculumOverview" | "subjectListingPage"
+  | "subjectPhaseOptions"
+  | "curriculumOverview"
+  | "subjectListingPage"
+  | "teachersHomePage"
 > = {
   subjectPhaseOptions: jest.fn(async () => {
     return subjectPhaseOptionsFixture();
@@ -15,6 +19,9 @@ const curriculumApi: Pick<
   }),
   subjectListingPage: jest.fn(async () => {
     return subjectListingFixture2023();
+  }),
+  teachersHomePage: jest.fn(async () => {
+    return teachersHomePageFixture();
   }),
 };
